News_summarizer/requirements.txt
streamlit==1.34.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
langchain==0.1.17
langchain-google-genai==0.0.6
